Separate test file at Cameron's request
this lets other tools easily use these tests
